Fix warnings in argparse.c

+++ TDEngine2/tools/ResourcesPacker/deps/argparse/argparse.c
@@ -279,7 +279,7 @@ unknown:
 
 end:
     memmove(self->out + self->cpidx, self->argv,
-            self->argc * sizeof(*self->out));
+            (size_t)(self->argc * sizeof(*self->out)));
     self->out[self->cpidx + self->argc] = NULL;
 
     return self->cpidx + self->argc;
